requirements.txt
kafka-python
easybase
python-dateutil
